getTempFileCopyPath is only used by tests so move to test-only code
