+++ common/FileUtil.cpp
@@ -55,32 +55,6 @@ namespace filesystem = ::std::filesystem;
 #include "Util.hpp"
 #include "Unit.hpp"
 
-namespace
-{
-#if HAVE_STD_FILESYSTEM
-/// Class to delete files when the process ends.
-class FileDeleter
-{
-    std::vector<std::string> _filesToDelete;
-    std::mutex _lock;
-public:
-    FileDeleter() {}
-    ~FileDeleter()
-    {
-        std::unique_lock<std::mutex> guard(_lock);
-        for (const std::string& file: _filesToDelete)
-            filesystem::remove(file);
-    }
-
-    void registerForDeletion(const std::string& file)
-    {
-        std::unique_lock<std::mutex> guard(_lock);
-        _filesToDelete.push_back(file);
-    }
-};
-#endif
-}
-
 namespace FileUtil
 {
     std::string createRandomDir(const std::string& path)
@@ -229,40 +203,6 @@ namespace FileUtil
             return root;
         }
         return newTmp;
-    }
-
-    std::string getTempFileCopyPath(const std::string& srcDir, const std::string& srcFilename, const std::string& dstFilenamePrefix)
-    {
-        const std::string srcPath = srcDir + '/' + srcFilename;
-#if HAVE_STD_FILESYSTEM
-        std::string dstPath;
-
-        bool retry;
-        do {
-            std::string dstFilename = dstFilenamePrefix + Util::encodeId(Util::rng::getNext()) + '_' + srcFilename;
-
-            retry = false;
-            dstPath = filesystem::temp_directory_path() / dstFilename;
-            try {
-                filesystem::copy(srcPath, dstPath);
-            }
-            catch (const std::exception& ex)
-            {
-                LOG_SYS("ERROR: unexpected conflict creating file: " << ex.what());
-                retry = true;;
-            }
-        } while (retry);
-
-        static FileDeleter fileDeleter;
-        fileDeleter.registerForDeletion(dstPath);
-#else
-        const std::string dstFilename = dstFilenamePrefix + Util::encodeId(Util::rng::getNext()) + '_' + srcFilename;
-        const std::string dstPath = Poco::Path(Poco::Path::temp(), dstFilename).toString();
-        copyFileTo(srcPath, dstPath);
-        Poco::TemporaryFile::registerForDeletion(dstPath);
-#endif
-
-        return dstPath;
     }
 
 #if 1 // !HAVE_STD_FILESYSTEM
